--- rust/adc_eth_test/src/macro_test.rs
@@ -1,7 +1,0 @@
-
-pub mod bazmod {
-    #[macro_export]
-    macro_rules! baz {
-        () => ()
-    }
-}
