+++ rust/temp_controller/README.md
