+++ rust/temp_controller/src/print.rs
